requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
langchain-google-genai==2.0.9
langchain_community==0.3.16
langchain==0.3.17
beautifulsoup4==4.13.3
pypdf==5.2.0
python-multipart==0.0.20
faiss-cpu==1.10.0
unstructured==0.11.8
pydantic_settings==2.7.1